+++ src/main/java/com/appdirect/sdk/vendorFields/controller/VendorRequiredFieldController.java
@@ -22,11 +22,11 @@ import com.appdirect.sdk.vendorFields.converter.FlowTypeConverter;
 import com.appdirect.sdk.vendorFields.converter.LocaleConverter;
 import com.appdirect.sdk.vendorFields.converter.OperationTypeConverter;
 import com.appdirect.sdk.vendorFields.handler.VendorRequiredFieldHandler;
-import com.appdirect.sdk.vendorFields.handler.VendorRequiredFieldHandlerV2;
+import com.appdirect.sdk.vendorFields.handler.v2.VendorRequiredFieldHandlerV2;
 import com.appdirect.sdk.vendorFields.model.FlowType;
-import com.appdirect.sdk.vendorFields.model.OperationType;
+import com.appdirect.sdk.vendorFields.model.v2.OperationType;
 import com.appdirect.sdk.vendorFields.model.VendorRequiredFieldsResponse;
-import com.appdirect.sdk.vendorFields.model.VendorRequiredFieldsResponseV2;
+import com.appdirect.sdk.vendorFields.model.v2.VendorRequiredFieldsResponseV2;
 
 /**
  * Defines the endpoint for enforcing vendor required requiredFields on their products
