+++ src/main/java/com/appdirect/sdk/vendorFields/converter/OperationTypeConverter.java
@@ -3,7 +3,7 @@ package com.appdirect.sdk.vendorFields.converter;
 import java.beans.PropertyEditorSupport;
 
 import com.appdirect.sdk.exception.PropertyEditorSupportException;
-import com.appdirect.sdk.vendorFields.model.OperationType;
+import com.appdirect.sdk.vendorFields.model.v2.OperationType;
 
 public class OperationTypeConverter extends PropertyEditorSupport {
 	public void setAsText(final String text) throws IllegalArgumentException {
