+++ src/main/java/com/appdirect/sdk/vendorFields/handler/v2/VendorRequiredFieldHandlerV2.java
@@ -1,11 +1,11 @@
-package com.appdirect.sdk.vendorFields.handler;
+package com.appdirect.sdk.vendorFields.handler.v2;
 
 import java.util.List;
 import java.util.Locale;
 
 import com.appdirect.sdk.vendorFields.model.FlowType;
-import com.appdirect.sdk.vendorFields.model.OperationType;
-import com.appdirect.sdk.vendorFields.model.VendorRequiredFieldsResponseV2;
+import com.appdirect.sdk.vendorFields.model.v2.OperationType;
+import com.appdirect.sdk.vendorFields.model.v2.VendorRequiredFieldsResponseV2;
 
 /**
  * This is the interface you need to implement to apply vendor required requiredFields on a product
