--- src/main/java/com/appdirect/sdk/vendorFields/model/Options.java
@@ -1,18 +1,0 @@
-package com.appdirect.sdk.vendorFields.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class Options {
-    private Suffix suffix;
-    private String placeholder;
-}
