+++ src/main/java/com/appdirect/sdk/vendorFields/model/VendorFieldsValidationRequest.java
@@ -8,6 +8,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import com.appdirect.sdk.vendorFields.model.v2.OperationType;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @Data
