+++ src/main/java/com/appdirect/sdk/vendorFields/model/v2/Suffix.java
@@ -1,4 +1,4 @@
-package com.appdirect.sdk.vendorFields.model;
+package com.appdirect.sdk.vendorFields.model.v2;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
