+++ src/main/java/com/appdirect/sdk/vendorFields/model/v2/ValidationFieldRequest.java
@@ -1,10 +1,11 @@
-package com.appdirect.sdk.vendorFields.model;
+package com.appdirect.sdk.vendorFields.model.v2;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import com.appdirect.sdk.vendorFields.model.v2.InputCode;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @Data
@@ -12,13 +13,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class VendorRequiredFieldV2 {
+public class ValidationFieldRequest {
     private InputCode inputCode;
-    private String inputTitle;
-    private String subTitle;
-    private String tooltip;
     private String value;
-    private FieldTypeV2 fieldType;
-    private Validations validations;
-    private Options options;
 }
