+++ src/main/java/com/appdirect/sdk/vendorFields/model/v2/VendorRequiredFieldsResponseV2.java
@@ -1,23 +1,18 @@
-package com.appdirect.sdk.vendorFields.model;
+package com.appdirect.sdk.vendorFields.model.v2;
 
 import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @Data
 @Builder
 @AllArgsConstructor
-@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-class FormV2 {
+public class VendorRequiredFieldsResponseV2 {
     private String isvIdentifier;
-    private Context context;
-    private String title;
-    private String subTitle;
     private List<VendorRequiredFieldV2> fields;
 }
